requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
pydantic-settings
aiofiles
google-generativeai
google-genai
opencv-python==4.8.1.78
mediapipe==0.10.9
numpy==1.26.2
Flask==3.0.0
Werkzeug==3.0.1
click==8.1.7
blinker==1.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
certifi